Extract per-class skill dispatch into GameLogic::Combat
Move the four per-class attack/skill executors (AttackElf, AttackKnight,
AttackWizard, AttackCommon, ~1,900 lines) out of ZzzInterface.cpp into
GameLogic/Combat/ClassAttack.{h,cpp} under namespace GameLogic::Combat.
ExecuteSkill (the only caller) now calls them qualified.

The functions still lean on file-scope state owned by ZzzInterface.cpp
(g_MovementSkill, the Target*/Selected* globals, etc.), declared via a small
extern block for now; those move once the shared skill/target state gets its
own header. GetDestValue stays in ZzzInterface.cpp with a new header decl.
No behaviour change.
